Updated Change Log for new updates
Updated a slew of mods, added things to the changelog, etc.
Make a a backup of the world, this could potentially break things due to the amount of content being added!

Hopefully, it doesn't though. :)
Note: There are quite a few mods being updated/added, so take your time when pushing this one out, test client compiling first before throwing it on the server.